+++ src/Engine/projects/project.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <string>
+
+class project
+{
+    std::string projectName;
+    std::string projectVersion;
+
+    void createNewProject();
+    void loadNewProject();
+};