+++ src/Engine/Engine.cpp
@@ -1,8 +1,11 @@
-#include "Engine.h"
 #include <iostream>
+#include <SDL2/SDL.h>
+
+#include "Engine.h"
 
 namespace ENGINE
 {
+
     engine::engine()
     {
         std::cout << "starting engine..." << std::endl; 
@@ -12,4 +15,5 @@ namespace ENGINE
     {
 
     }
+
 }