+++ src/Engine/Settings.cpp
@@ -28,6 +28,7 @@ namespace ENGINE
         try
         {
             ifs.open("settings.cfg");
+            
             if (ifs.is_open())
             {
                 throw ENGINE::ErrorResult(FILE_NOT_FOUND_CORRUPT,"settings.cfg");
@@ -39,12 +40,30 @@ namespace ENGINE
                 
             }
 
+        while(getline(ifs,line)) 
+        {
+            std::cout << line << std::endl;
+            settingsFileContents.push_back(line);
+        }
+        
+        
+        for(int i = 0; i < settingsFileContents.size(); i++)
+        {
+            for(int x = 0; properties.size() < x; x++)
+            {
+                //TODO: search for setting's keywords and then assign them
+            }
+        }
+
+
             
         }
         catch(const std::exception& e)
         {
             std::cerr << e.what() << '\n';
         }
+
+
         
     }