+++ src/Engine/Settings.h
@@ -10,6 +10,7 @@
 
 
 
+
 using std::string;
 
 namespace ENGINE
@@ -33,9 +34,29 @@ namespace ENGINE
 
             
         private:
+        std::vector <string> properties = {"debugKey","resX","resY","isFullscreen","volume_Master","volume_Music","volume_UI","volume_SFX"}; //settings variable names 
+
         ERRORDEF err;
         std::ifstream ifs;
+        std::vector <string> settingsFileContents;  
+        string line; 
 
+        //----Settings--Variables----//
+        int debugKey; //optional
+        
+        //--Video--//
+
+        int resX;     // screen resolution X
+        int resY;     // screen resolution Y
+        bool isFullscreen; //fullscreen ON/OFF
+
+        //--Audio--//
+        int volume_Master;
+        int volume_Music;
+        int volume_UI;
+        int volume_SFX;
+
+